+++ Scripts/Chocolatey/tools/chocolateyinstall.ps1
@@ -1,4 +1,4 @@
-﻿<#
+<#
 	.SYNOPSIS
 	Get direct URL of Sophia Script archive, depending on which Windows it is run on
 
@@ -16,71 +16,34 @@
 $Parameters = @{
 	Uri             = "https://api.github.com/repos/farag2/Sophia-Script-for-Windows/releases/latest"
 	UseBasicParsing = $true
+	Verbose         = $true
 }
 $LatestGitHubRelease = (Invoke-RestMethod @Parameters).tag_name
 
 $Parameters = @{
 	Uri             = "https://raw.githubusercontent.com/farag2/Sophia-Script-for-Windows/master/sophia_script_versions.json"
 	UseBasicParsing = $true
+	Verbose         = $true
 }
 $JSONVersions = Invoke-RestMethod @Parameters
 
-$null = $packageParameters
 $packageParameters = $env:chocolateyPackageParameters
 
 switch ((Get-CimInstance -ClassName Win32_OperatingSystem).BuildNumber)
 {
 	"17763"
 	{
-		# Check for Windows 10 LTSC 2019
-		if ((Get-ItemPropertyValue -Path "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion" -Name ProductName) -match "LTSC 2019")
-		{
-			$LatestRelease = $JSONVersions.Sophia_Script_Windows_10_LTSC2019
-			$URL = "https://github.com/farag2/Sophia-Script-for-Windows/releases/download/$LatestGitHubRelease/Sophia.Script.for.Windows.10.LTSC.2019.v$LatestRelease.zip"
-			$Hash = "Hash_Sophia_Script_Windows_10_LTSC2019"
-		}
-		else
-		{
-			Write-Verbose -Message "Windows version is not supported. Update your Windows" -Verbose
-
-			# Receive updates for other Microsoft products when you update Windows
-			(New-Object -ComObject Microsoft.Update.ServiceManager).AddService2("7971f918-a847-4430-9279-4a52d1efe18d", 7, "")
-
-			# Check for updates
-			Start-Process -FilePath "$env:SystemRoot\System32\UsoClient.exe" -ArgumentList StartInteractiveScan
-
-			# Open the "Windows Update" page
-			Start-Process -FilePath "ms-settings:windowsupdate"
-
-			pause
-			exit
-		}
+		# Windows 10 LTSC 2019
+		$LatestRelease = $JSONVersions.Sophia_Script_Windows_10_LTSC2019
+		$URL = "https://github.com/farag2/Sophia-Script-for-Windows/releases/download/$LatestGitHubRelease/Sophia.Script.for.Windows.10.LTSC.2019.v$LatestRelease.zip"
+		$Hash = "Hash_Sophia_Script_Windows_10_LTSC2019"
 	}
 	"19044"
 	{
-		# Check for Windows 10 LTSC 2021
-		if ((Get-ItemPropertyValue -Path "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion" -Name ProductName) -match "LTSC 2021")
-		{
-			$LatestRelease = $JSONVersions.Sophia_Script_Windows_10_LTSC2021
-			$URL = "https://github.com/farag2/Sophia-Script-for-Windows/releases/download/$LatestGitHubRelease/Sophia.Script.for.Windows.10.LTSC.2021.v$LatestRelease.zip"
-			$Hash = "Hash_Sophia_Script_Windows_10_LTSC2021"
-		}
-		else
-		{
-			Write-Verbose -Message "Windows version is not supported. Update your Windows" -Verbose
-
-			# Receive updates for other Microsoft products when you update Windows
-			(New-Object -ComObject Microsoft.Update.ServiceManager).AddService2("7971f918-a847-4430-9279-4a52d1efe18d", 7, "")
-
-			# Check for updates
-			Start-Process -FilePath "$env:SystemRoot\System32\UsoClient.exe" -ArgumentList StartInteractiveScan
-
-			# Open the "Windows Update" page
-			Start-Process -FilePath "ms-settings:windowsupdate"
-
-			pause
-			exit
-		}
+		# Windows 10 LTSC 2021
+		$LatestRelease = $JSONVersions.Sophia_Script_Windows_10_LTSC2021
+		$URL = "https://github.com/farag2/Sophia-Script-for-Windows/releases/download/$LatestGitHubRelease/Sophia.Script.for.Windows.10.LTSC.2021.v$LatestRelease.zip"
+		$Hash = "Hash_Sophia_Script_Windows_10_LTSC2021"
 	}
 	"19045"
 	{
@@ -117,7 +80,7 @@ switch ((Get-CimInstance -ClassName Win32_OperatingSystem).BuildNumber)
 				{
 					if ((Get-CimInstance -ClassName CIM_Processor).Caption -match "ARM")
 					{
-						$LatestRelease = $JSONVersions.Sophia_Script_Windows_11_Arm__PowerShell_7
+						$LatestRelease = $JSONVersions.Sophia_Script_Windows_11_Arm_PowerShell_7
 						$URL = "https://github.com/farag2/Sophia-Script-for-Windows/releases/download/$LatestGitHubRelease/Sophia.Script.for.Windows.11.Arm.PowerShell.7.v$LatestRelease.zip"
 						$Hash = "Hash_Sophia_Script_Windows_11_Arm_PowerShell_7"
 					}
@@ -135,26 +98,28 @@ switch ((Get-CimInstance -ClassName Win32_OperatingSystem).BuildNumber)
 				{
 					$LatestRelease = $JSONVersions.Sophia_Script_Windows_11_Arm_PowerShell_5_1
 					$URL = "https://github.com/farag2/Sophia-Script-for-Windows/releases/download/$LatestGitHubRelease/Sophia.Script.for.Windows.11.Arm.v$LatestRelease.zip"
-					$Hash = "Hash_Sophia_Script_Windows_11_Arm"
+					$Hash = "Hash_Sophia_Script_Windows_11_Arm_PowerShell_5_1"
 				}
 				else
 				{
 					$LatestRelease = $JSONVersions.Sophia_Script_Windows_11_PowerShell_5_1
 					$URL = "https://github.com/farag2/Sophia-Script-for-Windows/releases/download/$LatestGitHubRelease/Sophia.Script.for.Windows.11.v$LatestRelease.zip"
-					$Hash = "Hash_Sophia_Script_Windows_11_PowerShell_7"
+					$Hash = "Hash_Sophia_Script_Windows_11_PowerShell_5_1"
 				}
 			}
 		}
 	}
 }
 
+Write-Verbose -Message "URL is $URL. Hash is $Hash" -Verbose
+
 $Downloads = Get-ItemPropertyValue -Path "HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\User Shell Folders" -Name "{374DE290-123F-4565-9164-39C4925E467B}"
 $Parameters = @{
 	packageName   = $env:ChocolateyPackageName
-	fileType      = "ZIP"
-	unzipLocation = $Downloads
-	url           = $URL
-	checksum      = $Hash
-	checksumType  = "sha256"
+	UnzipLocation = $Downloads
+	Url           = $URL
+	Checksum      = $Hash
+	ChecksumType  = "sha256"
+	Verbose       = $true
 }
 Install-ChocolateyZipPackage @Parameters