+++ src/Sophia_Script_for_Windows_10_LTSC_2021/Import-TabCompletion.ps1
@@ -3,10 +3,10 @@
 	Enable tab completion to invoke for functions if you do not know function name
 
 	.VERSION
-	6.1.1
+	6.1.4
 
 	.DATE
-	13.02.2026
+	24.02.2026
 
 	.COPYRIGHT
 	(c) 2014—2026 Team Sophia